--- src/context.ts
@@ -1,11 +1,0 @@
-import { PrismaClient } from '@prisma/client'
-
-const prisma = new PrismaClient()
-
-export interface Context {
-  prisma: PrismaClient
-}
-
-export function createContext(): Context {
-  return { prisma }
-}
