+++ src/index.ts
@@ -1,8 +1,13 @@
 import { GraphQLServer } from 'graphql-yoga'
-import { createContext } from './context'
+import { permissions } from './permissions'
 import { schema } from './schema'
+import { createContext } from './utils'
 
-new GraphQLServer({ schema, context: createContext }).start(() =>
+new GraphQLServer({
+  schema,
+  context: createContext,
+  middlewares: [permissions],
+}).start(() =>
   console.log(
     `🚀 Server ready at: http://localhost:4000\n⭐️ See sample queries: http://pris.ly/e/ts/graphql-sdl-first#using-the-graphql-api`,
   ),
